requirements.txt
streamlit
opencv-python
pillow
numpy
easyocr